Change data type of llr from double to float in callers of run_ldpc_decoder.
Also change data type in Demod2D, Somap, and symbols_to_llts which setup data
for run_ldpc_decoder.

This causes some higher BER values for the fast fading test!

+++ codec2-dev/src/mpdecode_core.h
@@ -36,11 +36,10 @@ void encode(struct LDPC *ldpc, unsigned char ibits[], unsigned char pbits[]);
 
 int run_ldpc_decoder(struct LDPC *ldpc, char out_char[], float input[], int *parityCheckCount);
 
-// Used for test programs.
 void sd_to_llr(float llr[], double sd[], int n);
-void Demod2D(double symbol_likelihood[], COMP r[], COMP S_matrix[], float EsNo, float fading[], float mean_amp, int number_symbols);
-void Somap(double bit_likelihood[], double symbol_likelihood[], int number_symbols);
-void symbols_to_llrs(double llr[], COMP rx_qpsk_symbols[], float rx_amps[], float EsNo, float mean_amp, int nsyms);
+void Demod2D(float symbol_likelihood[], COMP r[], COMP S_matrix[], float EsNo, float fading[], float mean_amp, int number_symbols);
+void Somap(float bit_likelihood[], float symbol_likelihood[], int number_symbols);
+void symbols_to_llrs(float llr[], COMP rx_qpsk_symbols[], float rx_amps[], float EsNo, float mean_amp, int nsyms);
 
 struct v_node {
   int degree;
